Make artist profile pictures pressable
