+++ packages/mobile/src/screens/profile-screen/ArtistRecommendations/ArtistRecommendations.tsx
@@ -7,7 +7,7 @@ import {
 } from 'audius-client/src/common/store/social/users/actions'
 import { makeGetRelatedArtists } from 'audius-client/src/common/store/ui/artist-recommendations/selectors'
 import { fetchRelatedArtists } from 'audius-client/src/common/store/ui/artist-recommendations/slice'
-import { View } from 'react-native'
+import { TouchableOpacity, View } from 'react-native'
 import { useEffectOnce } from 'react-use'
 
 import IconFollow from 'app/assets/images/iconFollow.svg'
@@ -16,6 +16,7 @@ import IconClose from 'app/assets/images/iconRemove.svg'
 import { Button, IconButton, Text } from 'app/components/core'
 import { ProfilePicture } from 'app/components/user'
 import { useDispatchWeb } from 'app/hooks/useDispatchWeb'
+import { useNavigation } from 'app/hooks/useNavigation'
 import { useSelectorWeb } from 'app/hooks/useSelectorWeb'
 import { makeStyles } from 'app/styles'
 import { track, make } from 'app/utils/analytics'
@@ -74,6 +75,7 @@ const getRelatedArtistIds = makeGetRelatedArtists()
 export const ArtistRecommendations = (props: ArtistRecommendationsProps) => {
   const { onClose } = props
   const styles = useStyles()
+  const navigation = useNavigation()
   const { user_id, name } = useSelectProfile(['user_id', 'name'])
 
   const dispatchWeb = useDispatchWeb()
@@ -115,11 +117,21 @@ export const ArtistRecommendations = (props: ArtistRecommendationsProps) => {
     })
   }, [suggestedArtists, isFollowingAllArtists, dispatchWeb])
 
+  const handlePressArtist = useCallback(
+    artist => () => {
+      navigation.push({
+        native: { screen: 'Profile', params: { handle: artist.handle } },
+        web: { route: `/${artist.handle}` }
+      })
+    },
+    [navigation]
+  )
+
   const suggestedArtistNames = suggestedArtists.slice(0, 3)
 
   return (
     <View pointerEvents='box-none' style={styles.root}>
-      <View style={styles.header}>
+      <View style={styles.header} pointerEvents='box-none'>
         <IconButton
           icon={IconClose}
           styles={{ root: styles.dismissButton, icon: styles.dismissIcon }}
@@ -132,18 +144,22 @@ export const ArtistRecommendations = (props: ArtistRecommendationsProps) => {
       </View>
       <View style={styles.suggestedArtistsPhotos}>
         {suggestedArtists.map(artist => (
-          <ProfilePicture
+          <TouchableOpacity
+            onPress={handlePressArtist(artist)}
             key={artist.user_id}
-            profile={artist}
-            style={styles.suggestedArtistPhoto}
-          />
+          >
+            <ProfilePicture
+              profile={artist}
+              style={styles.suggestedArtistPhoto}
+            />
+          </TouchableOpacity>
         ))}
       </View>
       <View style={styles.suggestedArtistsText}>
         <Text variant='body1'>Featuring </Text>
         {suggestedArtistNames.map(artist => (
           <Fragment key={artist.user_id}>
-            <ArtistLink artist={artist} />
+            <ArtistLink artist={artist} onPress={handlePressArtist(artist)} />
             <Text variant='body1'>, </Text>
           </Fragment>
         ))}
